Tried to improve stop of nats server

+++ core/src/infra/natsserver.ts
@@ -45,17 +45,21 @@ export class NatsServer {
     protected clusterSeedUrls?: string[];
     protected serverListenPort?: number;
     protected clusterListenUrl?: string;
+    protected exeName?: string;
+    protected appName?: string;
 
     constructor(
         onUnexpectedStop?: (stderr: string) => void,
         serverListenPort?: number,
         clusterSeedUrls?: string[],
-        clusterListenUrl?: string
+        clusterListenUrl?: string,
+        appName?: string
     ) {
         this.onUnexpectedStop = onUnexpectedStop;
         this.clusterSeedUrls = clusterSeedUrls;
         this.serverListenPort = serverListenPort;
         this.clusterListenUrl = clusterListenUrl;
+        this.appName = appName;
     }
 
     public start(): void {
@@ -65,15 +69,22 @@ export class NatsServer {
         const filename = `nats-server-${platform}-${arch}${ext}`;
         const path = __dirname + '/../../binaries/';
         const fullname = path + filename;
+        this.exeName = fullname;
 
         const args = [];
         args.push('--cluster_name');
         args.push('darlean');
 
+        if (this.appName) {
+            args.push('--pid');
+            args.push(this.appName + '.pid');
+        }
+
         if (this.serverListenPort !== undefined) {
             args.push('--port');
             args.push(this.serverListenPort.toString());
         }
+
         if (this.clusterListenUrl) {
             args.push('--cluster');
             args.push(this.clusterListenUrl);
@@ -86,9 +97,10 @@ export class NatsServer {
 
         // console.log('STARTNG', fullname, args);
 
-        const process = cp.execFile(fullname, args, {}, (_error, _stdout, stderr) => {
+        const process = cp.execFile(fullname, args, (error, stdout, stderr) => {
             this.running = false;
             this.process = undefined;
+            // console.log('NATS', error, stdout, stderr);
             if (!this.stopping) {
                 this.onUnexpectedStop?.(stderr);
             }
@@ -98,8 +110,21 @@ export class NatsServer {
     }
 
     public stop(): void {
-        if (this.process) {
+        if (this.process && this.exeName && this.process.pid) {
             this.stopping = true;
+
+            // TODO: Improve shuwdown on windows. Wndows does not know signals, it just has a hard kill, but
+            // we want de graceful shutdown. We have filed an issue as Nats for this:
+            // https://github.com/nats-io/nats-server/issues/3809
+
+            // Note: ctrlc-windows package kills entire process tree including ourselves, so that
+            // does not work.
+
+            //const args = ['--signal', `stop=${this.process.pid}`]
+            //console.log('----- STOPPING NATS', this.exeName, args);
+            //cp.execFileSync(this.exeName, args);
+            //this.process.stdin?.write('\u001a');
+            //ctrlc.ctrlc(this.process.pid);
             this.process.kill();
             this.process = undefined;
         }
